Add experimental event based render graph This adds: - SPV reflection via shaderc - Resource management via the new events API from Vulkan 1.3 - Dynamic rendering Note that only the Path integrator works at the moment.

+++ src/Framework/MitsubaParser.cpp
@@ -15,7 +15,7 @@ void MitsubaParser::parse(const std::string& path) {
 				integrator.type = obj->pluginType();
 				for (const auto& prop : obj->properties()) {
 					if (prop.first == "max_depth") {
-						integrator.depth = prop.second.getInteger();
+						integrator.depth = (int)prop.second.getInteger();
 					} 
 					if (prop.first == "enable_vm") {
 						integrator.enable_vm = prop.second.getBool();
